Plan A: GRU HH/IH weights int16 Q12 → int32 Q20
extract_weights.m:
- GRU HH/IH weights reloaded from float (para_in_mat/) and
  re-quantized to int32 Q20 (×1048576)
- GRU Qr macros: Qr1 -13→-21, Qr2 -8→-16

MATLAB modules (21 files):
- All GRU_module calls: -13,-8 → -21,-16
- diag scripts updated

C code:
- gru_step_fp_q20: int16_t*→int32_t* for ih_weight/hh_weight
- gru_sequence_fp_q20, bigru_sequence_fp_q20: same
- ulunas_fp.h: all Q20 GRU declarations updated
- Hardcoded GRU Qr values: -13→-21, -8→-16

Weight precision: Q12 (12-bit) → Q20 (20-bit) = +8 bits
Effective MAC: Q20×Q20=Q40, >>21=Q19 (same result, 8× finer weight)

+++ x2000_deploy_v2/ulunas_fp.h
@@ -196,8 +196,8 @@ void affineprelu_fp(
 void gru_step_fp(
     const int32_t *x_t, int input_dim,
     int16_t *h_cache, int nHidden,
-    const int16_t *ih_weight, const int32_t *ih_bias,
-    const int16_t *hh_weight, const int32_t *hh_bias,
+    const int32_t *ih_weight, const int32_t *ih_bias,
+    const int32_t *hh_weight, const int32_t *hh_bias,
     int Qr1, int Qr2,
     int16_t *y_out);
 
@@ -205,8 +205,8 @@ void gru_step_fp(
 void gru_sequence_fp(
     const int32_t *x, int T, int input_dim,
     int16_t *h_cache, int nHidden,
-    const int16_t *ih_weight, const int32_t *ih_bias,
-    const int16_t *hh_weight, const int32_t *hh_bias,
+    const int32_t *ih_weight, const int32_t *ih_bias,
+    const int32_t *hh_weight, const int32_t *hh_bias,
     int Qr1, int Qr2,
     int16_t *y_out);
 
@@ -214,8 +214,8 @@ void gru_sequence_fp(
 void bigru_sequence_fp(
     const int32_t *x, int T, int input_dim,
     int nHidden,
-    const int16_t *ih_weight, const int32_t *ih_bias,
-    const int16_t *hh_weight, const int32_t *hh_bias,
+    const int32_t *ih_weight, const int32_t *ih_bias,
+    const int32_t *hh_weight, const int32_t *hh_bias,
     const int16_t *re_ih_weight, const int32_t *re_ih_bias,
     const int16_t *re_hh_weight, const int32_t *re_hh_bias,
     int Qr1, int Qr2,
@@ -225,24 +225,24 @@ void bigru_sequence_fp(
 void gru_step_fp_q20(
     const int32_t *x_t, int input_dim,
     int32_t *h_cache, int nHidden,
-    const int16_t *ih_weight, const int32_t *ih_bias,
-    const int16_t *hh_weight, const int32_t *hh_bias,
+    const int32_t *ih_weight, const int32_t *ih_bias,
+    const int32_t *hh_weight, const int32_t *hh_bias,
     int Qr1, int Qr2,
     int16_t *y_out);
 
 void gru_sequence_fp_q20(
     const int32_t *x, int T, int input_dim,
     int32_t *h_cache, int nHidden,
-    const int16_t *ih_weight, const int32_t *ih_bias,
-    const int16_t *hh_weight, const int32_t *hh_bias,
+    const int32_t *ih_weight, const int32_t *ih_bias,
+    const int32_t *hh_weight, const int32_t *hh_bias,
     int Qr1, int Qr2,
     int16_t *y_out);
 
 void bigru_sequence_fp_q20(
     const int32_t *x, int T, int input_dim,
     int nHidden,
-    const int16_t *ih_weight, const int32_t *ih_bias,
-    const int16_t *hh_weight, const int32_t *hh_bias,
+    const int32_t *ih_weight, const int32_t *ih_bias,
+    const int32_t *hh_weight, const int32_t *hh_bias,
     const int16_t *re_ih_weight, const int32_t *re_ih_bias,
     const int16_t *re_hh_weight, const int32_t *re_hh_bias,
     int Qr1, int Qr2,
@@ -287,8 +287,8 @@ void mask_fp(
 void ctfa_ta_fp(
     const int32_t *x, int C, int W,
     int32_t *ta_h_cache, int nHidden, int input_dim,
-    const int16_t *ih_weight, const int32_t *ih_bias,
-    const int16_t *hh_weight, const int32_t *hh_bias,
+    const int32_t *ih_weight, const int32_t *ih_bias,
+    const int32_t *hh_weight, const int32_t *hh_bias,
     const int16_t *fc_weight, const int32_t *fc_bias,
     uint16_t *ta_out);
 
@@ -296,8 +296,8 @@ void ctfa_ta_fp(
 void ctfa_fa_fp(
     const int32_t *x, int C, int W, int pad_len,
     int nHidden,
-    const int16_t *ih_weight, const int32_t *ih_bias,
-    const int16_t *hh_weight, const int32_t *hh_bias,
+    const int32_t *ih_weight, const int32_t *ih_bias,
+    const int32_t *hh_weight, const int32_t *hh_bias,
     const int16_t *re_ih_weight, const int32_t *re_ih_bias,
     const int16_t *re_hh_weight, const int32_t *re_hh_bias,
     const int16_t *fc_weight, const int32_t *fc_bias,
